fix archive.cpp on 32-bit

+++ src/support/archive.cpp
@@ -211,7 +211,7 @@ static uint32_t read32be(const uint8_t* buf) {
 
 void Archive::dump() const {
   printf("Archive data %p len %lu, firstRegularData %p\n", data.data(),
-         data.size(), firstRegularData);
+         (long unsigned)data.size(), firstRegularData);
   printf("Symbol table %p, len %u\n", symbolTable.data, symbolTable.len);
   printf("string table %p, len %u\n", stringTable.data, stringTable.len);
   const uint8_t* buf = symbolTable.data;
